requirements.txt
# MLX and Whisper dependencies
mlx-whisper>=0.4.3
mlx>=0.29.1

# Audio processing (requires PortAudio: brew install portaudio)
pyaudio>=0.2.14
numpy>=2.3.0

# Menu bar interface
rumps>=0.4.0

# Global hotkey detection
pynput>=1.8.1

# macOS system integration
pyobjc-framework-Cocoa>=11.1
pyobjc-framework-Quartz>=11.1
pyobjc-framework-ApplicationServices>=11.1

# Configuration and data handling
pyyaml>=6.0

# Additional dependencies for MLX Whisper
torch>=2.8.0
tqdm>=4.67.0
huggingface-hub>=0.35.0

